Move parameter utils to Utils module
